+++ src/simulation/user/params.rs
@@ -3,5 +3,5 @@ use rune::Any;
 #[derive(Any)]
 pub struct UserParams {
     #[rune(get)]
-    pub user_id: u32,
+    pub user_id: u64,
 }
